refactor(inputs): new hierarchy - PureInput -> Input
