+++ packages/input/src/Component.module.css
@@ -1,8 +1,10 @@
 @import '../../themes/default.css';
+
 :root {
-  --input-font-family: var(--font-family);
   --input-hint-color: var(--color-dark-indigo-60-flat);
   --input-error-color: var(--color-red-brand-85-flat);
+  --input-sub-font-size: 14px;
+  --input-sub-line-height: 18px;
 }
 
 .component {
@@ -11,18 +13,100 @@
   box-sizing: border-box;
 }
 
-.left-icon {
-  margin-left: var(--gap-s);
+.component:not(.disabled) .inner:hover {
+  background-color: var(--input-hover-bg-color);
 }
 
-.right-icon {
-  margin-right: var(--gap-s);
+.disabled.focused .inner {
+  border-bottom-color: var(--input-focus-border-color);
+  box-shadow: inset 0 -1px var(--input-focus-border-color);
+}
+
+.disabled .inner {
+  border-bottom: 1px dashed var(--input-disabled-border-color);
+  background-color: var(--input-disabled-bg-color);
+}
+
+.disabled .input {
+  color: var(--input-disabled-color);
+}
+
+.inner {
+  display: flex;
+  position: relative;
+  background-color: var(--input-bg-color);
+  border-radius: var(--input-border-radius);
+  border-bottom: 1px solid var(--input-border-color);
+  transition: background 0.2s ease, box-shadow 0.2s ease, border 0.2s ease;
+}
+
+.input {
+  position: relative;
+  width: 100%;
+  color: var(--input-color);
+  background: none;
+  border: none;
+  margin: 0;
+  outline: none;
+  transition: background 0.2s ease;
+  font-size: var(--input-font-size);
+  line-height: var(--input-line-height);
+  box-sizing: border-box;
+  -webkit-appearance: none;
+  -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+}
+
+.input::placeholder,
+.input[placeholder] {
+  text-overflow: ellipsis;
+  transition: opacity 0.2s ease-in-out, color 0.2s ease-in-out;
+}
+
+.input::placeholder {
+  color: var(--input-placeholder-color);
+}
+
+.input-wrapper {
+  flex-grow: 1;
+  position: relative;
+}
+
+.label {
+  overflow: hidden;
+  position: absolute;
+  display: flex;
+  align-items: center;
+  height: 100%;
+  left: var(--input-side-paddings);
+  color: var(--input-placeholder-color);
+  font-size: var(--input-font-size);
+  line-height: var(--input-line-height);
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  transform: scale(1) translateY(0);
+  transform-origin: 0 100%;
+  transition-duration: 200ms;
+  transition-property: color, transform, font-size;
+  transition-timing-function: cubic-bezier(0.25, 0.1, 0.25, 1);
+}
+
+.addons {
+  display: flex;
+  flex-shrink: 0;
+  align-items: center;
+}
+.addons:last-child {
+    padding-right: var(--gap-s);
+}
+
+.addons:first-child {
+  padding-right: var(--gap-s);
 }
 
 .sub {
   display: block;
-  font-size: 14px;
-  line-height: 18px;
+  font-size: var(--input-sub-font-size);
+  line-height: var(--input-sub-line-height);
   margin-top: var(--gap-2xs);
   color: var(--input-hint-color);
 }
@@ -34,4 +118,43 @@
 .has-error .inner {
   box-shadow: inset 0 -1px var(--input-error-color);
   border-bottom-color: var(--input-error-color);
+}
+
+.s {
+  padding: 14px var(--input-side-paddings) 13px var(--input-side-paddings);
+}
+
+.m {
+  padding: 18px var(--input-side-paddings) 17px var(--input-side-paddings);
+}
+
+.l {
+  padding: 24px var(--input-side-paddings) 23px var(--input-side-paddings);
+}
+
+.has-label .input::placeholder {
+  opacity: 0;
+}
+
+.has-label.focused .input::placeholder {
+  opacity: 1;
+}
+
+.focused .label,
+.filled .label {
+  font-size: 14px;
+  line-height: 18px;
+  transform: scale(1) translateY(-10px);
+}
+
+.has-label .s {
+  padding: 23px var(--input-side-paddings) 4px var(--input-side-paddings);
+}
+
+.has-label .m {
+  padding: 28px var(--input-side-paddings) 7px var(--input-side-paddings);
+}
+
+.has-label .l {
+  padding: 34px var(--input-side-paddings) 13px var(--input-side-paddings);
 }
