+++ packages/input/src/Component.stories.tsx
@@ -14,7 +14,13 @@ import { withDesign } from 'storybook-addon-designs';
 
 import { Input } from './Component';
 
-// TODO: move icons
+export default {
+  title: 'Common',
+  component: Input,
+  decorators: [withDesign, withKnobs]
+};
+
+// TODO: Этого не будет, когда появится компонент иконки.
 const icon = (
   <svg width='18' height='18' viewBox='0 0 18 18' fill='none' xmlns='http://www.w3.org/2000/svg'>
     <path fillRule='evenodd' clipRule='evenodd' d='M9 1C4.588 1 1 4.588 1 9C1 13.412 4.588 17 9 17C13.412 17 17 13.412 17 9C17 4.588 13.412 1 9 1ZM9 15.933C7.1619 15.9309 5.3997 15.1998 4.09997 13.9C2.80024 12.6003 2.06912 10.8381 2.067 9C2.06912 7.1619 2.80024 5.3997 4.09997 4.09997C5.3997 2.80024 7.1619 2.06912 9 2.067C10.8381 2.06912 12.6003 2.80024 13.9 4.09997C15.1998 5.3997 15.9309 7.1619 15.933 9C15.9309 10.8381 15.1998 12.6003 13.9 13.9C12.6003 15.1998 10.8381 15.9309 9 15.933Z' fill='#0B1F35' />
@@ -22,56 +28,24 @@ const icon = (
   </svg>
 );
 
-export default {
-  title: 'Common',
-  component: Input,
-  decorators: [withDesign, withKnobs]
-};
-
 export const InputStory = () => {
   const [value, setValue] = useState('value');
 
   const handleChange = (event: React.ChangeEvent<HTMLInputElement>): void => setValue(event.target.value);
 
   return (
-    <div>
-      <div style={ { marginBottom: '20px' } }>
-        <Input
-          size={ select('size', ['s', 'm', 'l'], 's') }
-          type={ select('type', ['text', 'number', 'card', 'email', 'file', 'hidden', 'money', 'password', 'tel', 'text'], 'text') }
-          disabled={ boolean('Disabled', false) }
-          placeholder={ text('placeholder', '') }
-          label={ text('label', '') }
-          hint={ text('hint', '') }
-          error={ text('error', '') }
-          value={ value }
-          onChange={ handleChange }
-        />
-      </div>
-
-      <div style={ { marginBottom: '20px' } }>
-        <Input
-          size='s'
-          placeholder='Enter text'
-          label='Search'
-          value={ value }
-          onChange={ handleChange }
-          leftIcon={ icon }
-        />
-      </div>
-
-      <div style={ { marginBottom: '20px' } }>
-        <Input
-          size='s'
-          placeholder='Enter text'
-          label='Search'
-          value={ value }
-          error='error'
-          onChange={ handleChange }
-          leftIcon={ icon }
-        />
-      </div>
-    </div>
+    <Input
+      type={ select('type', ['text', 'number', 'card', 'email', 'file', 'hidden', 'money', 'password', 'tel', 'text'], 'text') }
+      size={ select('size', ['s', 'm', 'l'], 's') }
+      disabled={ boolean('disabled', false) }
+      placeholder={ text('placeholder', '') }
+      label={ text('label', '') }
+      hint={ text('hint', '') }
+      error={ text('error', '') }
+      value={ value }
+      onChange={ handleChange }
+      rightAddons={ boolean('withIcon', true) && (!text('error', '') && icon) }
+    />
   );
 };
