+++ packages/pure-input/src/Component.module.css
@@ -1,0 +1,75 @@
+@import '../../themes/default.css';
+
+:root {
+  --input-font-family: var(--font-family);
+  --input-font-size: 16px;
+  --input-line-height: 16px;
+  --input-color: var(--color-dark-indigo);
+  --input-disabled-color: var(--color-dark-indigo-60-flat);
+  --input-disabled-bg-color: var(--color-dark-indigo-15);
+  --input-bg-color: var(--color-dark-indigo-07);
+  --input-hover-bg-color: var(--color-dark-indigo-15);
+  --input-placeholder-color: var(--color-dark-indigo-60-flat);
+  --input-border-color: var(--color-dark-indigo-30-flat);
+  --input-focus-border-color: var(--color-dark-indigo);
+  --input-disabled-border-color: var(--color-dark-indigo-30-flat);
+  --input-border-radius: 4px 4px 0 0;
+  --input-side-paddings: 12px;
+}
+
+.component {
+  display: flex;
+  position: relative;
+  width: 100%;
+  max-width: 310px;
+  font-family: var(--input-font-family);
+  font-size: var(--input-font-size);
+  line-height: var(--input-line-height);
+  box-sizing: border-box;
+  color: var(--input-color);
+  background-color: var(--input-bg-color);
+  border-radius: var(--input-border-radius);
+  transition: background 0.2s ease, box-shadow 0.2s ease, border 0.2s ease, opacity 0.2s ease-in-out, color 0.2s ease-in-out;
+  border: none;
+  margin: 0;
+  border-bottom: 1px solid var(--input-border-color);
+  outline: none;
+  -webkit-appearance: none;
+  -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+}
+
+.component::placeholder,
+.component[placeholder] {
+  text-overflow: ellipsis;
+}
+
+.component::placeholder {
+  color: var(--input-placeholder-color);
+}
+
+.component:not(.disabled):hover {
+  background-color: var(--input-hover-bg-color);
+}
+
+.component.disabled:focus {
+  border-bottom-color: var(--input-focus-border-color);
+  box-shadow: inset 0 -1px var(--input-focus-border-color);
+}
+
+.disabled {
+  color: var(--input-disabled-color);
+  border-bottom: 1px dashed var(--input-disabled-border-color);
+  background-color: var(--input-disabled-bg-color);
+}
+
+.s {
+  padding: 14px var(--input-side-paddings) 13px var(--input-side-paddings);
+}
+
+.m {
+  padding: 18px var(--input-side-paddings) 17px var(--input-side-paddings);
+}
+
+.l {
+  padding: 24px var(--input-side-paddings) 23px var(--input-side-paddings);
+}
